requirements.txt
Pillow==10.0.0
tkinter
pytsk3
subprocess
threading